\name{Grho-package}
\alias{Grho-package}
%\alias{Grho}
\docType{package}
\title{
Grho parameters for asymptotic regression
}
\description{
  For NDF assay data it is more convenient to parameterize the
  asymptotic regression model in terms of the expected values at three
  time points.  This package provides evaluate the fitted values from
  the expected value parameters corresponding to three specific times  -
  24, 36 and 48 hours.
}
\details{
\tabular{ll}{
Package: \tab Grho\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2008-07-29\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
  Douglas Bates and John Goeser.

  Maintainer: Douglas Bates <bates@R-project.org>
}
%\references{}
\keyword{ package }
%\seealso{}
%\examples{}
